requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
pillow
pyautogui
opencv-python
psutil
pynput
keyboard
mouse
pycaw
pywin32
psutil
